+++ test/Braintree.Tests/AddressTest.cs
@@ -1,6 +1,7 @@
 using Braintree.Exceptions;
 using Newtonsoft.Json;
 using NUnit.Framework;
+using System.Globalization;
 using System.Text;
 using System.Xml;
 
@@ -80,8 +81,10 @@ namespace Braintree.Tests
             Assert.AreEqual("1", address.CountryCodeNumeric);
             Assert.AreEqual("United States", address.CountryName);
             Assert.AreEqual("555-555-5555", address.PhoneNumber);
-            Assert.AreEqual("10/10/2018 22:46:41", address.CreatedAt.ToString());
-            Assert.AreEqual("10/10/2020 22:46:41", address.UpdatedAt.ToString());
+            Assert.AreEqual("10/10/2018 10:46:41 PM",
+                    address.CreatedAt?.ToString(CultureInfo.GetCultureInfo("en-US")));
+            Assert.AreEqual("10/10/2020 10:46:41 PM",
+                    address.UpdatedAt?.ToString(CultureInfo.GetCultureInfo("en-US")));
         }
     }
 }